add new file valve.rs and refactor
